Added a way to handle no cards in a deck, and a way to delete decks and cards
Allows users to click on decks with no cards. If a deck has no cards the user will be prompted with a button that when clicked, brings them to the create card screen.

Also allow users to delete decks and cards. Users can click a delete button to be prompted into a delete mode. When in this delete mode, if a user clicks on an item in the list, that item will be chosen to be deleted and the user will be asked to confirm their decision. If they are, the item will be deleted and the updated list will be presented to the user